Adapt dates used in test 10 to be relative to now
Since some weeks, squids/t/10-regression-mingle-356-bugzilla-46269.t
failed, as the used hard-coded dates were older than a year, which is
above the threshold allowed to run scripts for. We now change the
hardcoded dates into dates relative to the current date (cf.
1a4cef4344601dd8c8e541ded76218fc6b7813ab), and thereby get the tests
to pass again.

In particular, the test generates data for the previous month (relative
to the current month) and it writes that data to disk. All other
operations and checks are performed on paths constructed from that date.

Bug: 63478
Change-Id: I5983e5897b4a75cbe9dcedb0d76143adb77f55f4